Revert "Revert "Put the tests in separate files""
This reverts commit bc8a2dfedb692897f4d6936932fcb29c54733596.
